pyproject.toml
[project]
name = "wolvesagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentscope>=1.0.5",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
]
